updated dependencies

+++ src/algorithms/mod.rs
@@ -8,8 +8,8 @@ pub fn random_new_population(population: &[DNA], new_population: &mut [DNA]) {
         .iter_mut()
         .take(population.len())
         .for_each(|individual| {
-            let parent_1 = &population[rand::random::<usize>() % population.len()];
-            let parent_2 = &population[rand::random::<usize>() % population.len()];
+            let parent_1 = &population[rand::random::<u32>() as usize % population.len()];
+            let parent_2 = &population[rand::random::<u32>() as usize % population.len()];
             *individual = parent_1.crossover(parent_2);
         });
 }
